% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH_LT10.R
\name{CH_LT10}
\alias{CH_LT10}
\title{Calculate 10 percent lethal temperatures}
\usage{
CH_LT10(Calculated_CU_FU10 = NULL)
}
\arguments{
\item{Calculated_CU_FU10}{a dataframe of calculated parameters}
}
\value{
the calculated 10 percent lethal temperatures
}
\description{
Variables used: Temp_max.lag1, CU_1119, FU_acc_log, FU_state
}
\details{
Calculates the 10 percent lethal temperatures for Sweetheart sweet cherries in the Okanagan Valley based on daily air temperatures.
}
