% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{cherrycoldhardiness}
\alias{cherrycoldhardiness}
\title{Shiny app for cold hardiness estimations}
\usage{
cherrycoldhardiness()
}
\value{
web application, data frame of lethal temperatures
}
\description{
Web document to calculate cold hardiness estimations
}
\details{
Calculates the 10. 50, and 90 percent lethal temperatures for Sweetheart sweet cherries in the Okanagan Valley based on daily air temperatures.
}
