% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CU_FU.R
\name{CU_FU}
\alias{CU_FU}
\title{Calculate chilling units and forcing units}
\usage{
CU_FU(data_input = NULL)
}
\arguments{
\item{data_input}{the user uploaded .csv file}
}
\value{
the calculated 90 percent lethal temperatures
}
\description{
Variables used: Temp_max.lag1, CU_1119, FU_acc_log, FU_state
}
\details{
Calculates chilling and focing units based on work by Neilsen et al. (2015)
}
