% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Accesses the data from URLs}
\usage{
getData(
  stations,
  folder,
  timeframe = c("hourly", "daily", "monthly"),
  verbose = TRUE,
  delete = TRUE
)
}
\value{
hourly temperature data
}
\description{
Retrieves hourly temperature data from the Government of Canada's Historic climate data
}
\details{
Variables used: ID, start, end

Retrieves hourly temperature data from the Government of Canada's Historic climate data
}
