% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH_LT50.R
\name{CH_LT50}
\alias{CH_LT50}
\title{Calculate 50 percent lethal temperatures}
\usage{
CH_LT50(Calculated_CU_FU = NULL)
}
\description{
Variables used: Temp_max.lag1, CU_1119, FU_acc_log, FU_state
}
\details{
Calculates the 50 percent lethal temperatures for Sweetheart sweet cherries in the Okanagan Valley based on daily air temperatures.
}
