% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH_LT90.R
\name{CH_LT90}
\alias{CH_LT90}
\title{Calculate 90 percent lethal temperatures}
\usage{
CH_LT90(Calculated_CU_FU = NULL)
}
\arguments{
\item{Calculated_CU_FU}{a dataframe of calculated parameters}

\item{Model90}{model for calculating 90 percent lethal temperatures}
}
\value{
the calculated 90 percent lethal temperatures
}
\description{
Variables used: Temp_max.lag1, CU_1119, FU_acc_log, FU_state
}
\details{
Calculates the 90 percent lethal temperatures for Sweetheart sweet cherries in the Okanagan Valley based on daily air temperatures.
}
