% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genURLS.R
\name{genURLS}
\alias{genURLS}
\title{Generates URLS to access Government of Canada's Historic climate data}
\usage{
genURLS(id, start, end, timeframe = c("hourly", "daily", "monthly"))
}
\value{
hourly temperature data
}
\description{
Retrieves hourly temperature data from the Government of Canada's Historic climate data
}
\details{
Variables used: ID, start, end

Retrieves hourly temperature data from the Government of Canada's Historic climate data
}
