% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Accesses data from URLs}
\usage{
getData(
  stations,
  folder,
  timeframe = c("hourly", "daily", "monthly"),
  verbose = TRUE,
  delete = TRUE
)
}
\value{
hourly temperature data
}
\description{
This function is used to get the temperature data from the generated URL to access the Gov. of Canada's Historic Climate data base. Modified code originally written by Gavin Simpson that can be found at: https://gist.github.com/gavinsimpson/8c13e3c5f905fd67cf85
}
\details{
Variables used: ID, start, end

Retrieves hourly temperature data from the Government of Canada's Historic climate data
}
